Release and acquire the GIL as needed.
The `*_cy_call_fct` will be called by the C++ code in a different thread.
To avoid race condition, we need to acquire the GIL before doing python
operation.

However, those function will also be called from the same thread
(in addArticle). To avoid deadlock we need to release the GIL when we
enter C++ function.