Acquire the GIL before decrementing the article ref.
If the ZimArticle is not referenced in the python world after the call
to add_article[1], the only reference to the zimArticle is the one the
wrapper has.

The wrapper is "stored" in a std::shared_ptr. When libzim ends using it,
it will decrement the reference in the std::shared_ptr, and if it comes
to 0, it will deleted the Wrapper, which in turn, will decrement the
ref to the article, and potentially makes the ZimArticle to be deleted.

However, everything is done in a different thread, so we must protect
the decrement by acquiring the GIL first.

[1] As this :
```
for i in range(N):
    article = ZimArticle(...)
    zimCreator.add_article(article)
```