Forward exceptions from Article's virtual methods
Upon adding an article, virtual methods get_xxx() are called on the article to retrieve data.
Should an exception be raised in this user code, it was caught and ignored.
This changes this behavior to raise a RuntimeError with an explicit message instead.
As this exception passes through Cython, we can't retrieve the actually raised exception though.