Added ability to choose compression algoritm
libzim 6.1.8 added the ability to choose the compression type on creation.
This adds support for it in the Creator and other required parts.

On user API, the main change is a `compression` argument to the Creator.
This can receive either a string (none, zip, lzma, zstd) or a member
of the new `Compression` enum (members named after strings above)

This enum converts to zim::CompressionType and that is forwarded down to the libzim.

I chose to add the new `compression` argument to the Creator before the `min_chunk_size`
as it seems there might better chances people would customize it than the chunks size.
I might be wrong or it might just not be wanted. Let me know and I'd change that
Because of this change, two tests using positional arguments were changed.