Update the reading part to the new libzim api.
The complex thing here is that libzim forbid the creation of an
empty/invalid `Entry` or `Item`.

However, cython generate the code
`cdef wrapper.Entry entry = foo()`
as :
```
zim::Entry entry;
entry = foo();
```

As there is no default constructor, this doesn't compile.
So we have to store a vector as this :
`cdef wrapper.Entry* entry = new Entry(foo())`
But, yet again, this is generated by cython as :
```
zim::Entry* entry;
zim::Entry tmp_value;
tmp_value = foo();
entry = new Entry(tmp_value);
```

The solution is to never have cython to handle the plain Entry and
all the time handle pointer.
So we have to create the Entry on the heap in the cpp world.
So we have to inherit `Entry`/`Item`/`Archive` in `lib.h` to make them
return a pointer instead of a plain value.