Removed duplicate libzim from dist packages
For dependencies resolution during compilation, we need to resolve `-l zim`, essentially
looking for libzim.[so|dylib]
Run-time looks for the versioned file instead (libzim.so.7 or libzim.7.dylib)
This is handled using a symlink when building but symlink are not supported in ZIP files
so those are added as whole files, making the archive a lot larger without any benefit.
Especially for linux where we have .7 and .7.0.0 files.

This removes the addl-all directive and replaces it with a list of expected patterns