WIP: Switch to Illustration API
Creator `set_faviconpath(path)` is now replaced with `add_illustration(size, content)`
which instead of recording the path of a Favicon Entry, receives the bytes of a PNG
image for a specific size (adding a 48 sized one being mandatory as per zim spec.)

Archive `has_favicon_entry` is replaced with `has_illustration(size=None)`
Archive `favicon_entry` is replaced with `get_illustration_item(size=None)`
Size is optionnal and defaults to 48 in libzim.
Note that we used to retrieve an Entry and now receive an Item.

Additionnaly, Archive gets a `get_illustration_sizes()` which returns a set of int sizes
for which an illustration is present.
WARNING: this is not implemented ATM and simply return a set with 48 regarless of what's
currently in the ZIM.