safer get_hints handling
Failsafe approach to Hints usage: only fail on missing or non-dict-like values.
Assume, `hints=` can be whatever and
 - only cherry-pick our expected Hint from there
 - cast any value to bool

While this drifts a bit appart the zero-logic approach of the wrapper, it's a common
pythonic approach for this kind of things.
Main reason for this is that filtering and casting in Python allows exception to
go up the stack while in-cpp casting does not and ends up ignored.

Now testing:
 - Item without `get_hints()`
 - writer.Item without `get_hints()` impl.
 - `get_hints()` returning a non-dict value
 - `get_hints()` returning unexpected hints (filtered)
 - hints on redirection