Make libzim a package.
It is possible to import libzim in python :
```
import libzim
libzim.reader.Archive("foo.zim")
```

Libzim is imported and submodule is created and correctly set.

But it is not possible to directly import `libzim.reader`.
Although the modules are created, the python import mechanism don't
recognize libzim as a package and fail to locate the submodule.

So we have to tweak a bit the system :
- Add a __path__ to libzim to "mark" it as a package.
- Provide our own module loader to not make python trying to load
  'libzim.writer` from `___LIBZIM___/writer.py`.

Now it is possible to directly import a submodule :
```
from libzim.reader import Archive
Archive("foo.zim")
```