The `CompressionType` is now a scoped enum `Compression`.
Cython stable don't support scoped enum. (Next version will).
For now we have to use a workaround.

We declare a `zim::Compression` enum but we never use it in python.
We convert from the python value (to int) using cython and
to c++ `zim::Compression` (from int) using a small helper function.