Use fullname instead of path in ModuleFinder
Hacking around the path for loading our modules is fragile as there are stuff
relying on its value to consider it a package.
`help(libzim)` for instance is broken as it can't enumerate it properly.

This is simpler, less hacky and fixes the module's docstring which is important
for self-discovery.