Supply mandatory metadata in writer example (#193)
* Supply mandatory metadata in writer example

A zim file has certain mandatory requirements in the metadata; for example, Language, Date, and Illustration must be present. If they aren't, then some ZIM reader apps (such as Kiwix for iOS) will refuse to open the zim file. So, add these mandatory pieces to the example code in the readme, so that if you generate a zim file by following the example code, it will be valid.

* Apply suggestions from code review

Co-authored-by: rgaudin <reg@rskg.org>

---------

Co-authored-by: rgaudin <reg@rskg.org>