Introducing Type Hints
As we are producing a binary extension via Cython, the regular source-based mechanism of type discovery is not available.
This makes using libzim a little less pleasant for those relying on types.

This adds type stubs that we will have to manually maintain.

I had to tweak a bit the wrapper to be able to expose the get_indexdata thing.
Previously, we were relying on the fact that the Item had no `get_indexdata` method to decide
whether to use the libzim auto-index feature (relied on heavily for HTML entries).

Conditional method is not really compatible with static typing.

In order to retain functionality and API, I chose to add a `get_indexdata` variable on all `Item`, set to `None`.
If this variable is `None` (no action taken), then we run the auto-index.
If it is not None but returns NULL, then no index (no auto neither)
If it is set and returns a proper IndexData, it is used.
If the variable is missing (not sub-classing `Item`), run the auto-index.

In other words, behavior stays the same but type checker have a `get_indexdata` variable to look-for.